Non-Signature Move Tweaks
I fixed a couple of typos with 2 signature moves but this patch is to update all none signature moves to be under 100bp. This is for the Signature attack updating im working on. 

Moves Changed
Shining Nova
Holy Judgment
Prominence Beam
Infinity Burn
Shadow Fall
Hide and Seek
Evil Squall